Reorganize the GameLoader, simplify and restructure.
